RavenDB-18189 abstract smuggler - validate options
